Change to use jekyll-assets helper
When the site is built with JEKYLL_ENV=production, jekyll-assets will mangle the paths to the assets with a hash, and any references to the assets needs to know about the path. Hence we now need to use the jekyll-assets provided helpers. e.g. https://github.com/envygeeks/jekyll-assets/tree/v2.2.8#tags